requirements.txt
mariadb
pycryptodome